+++ console
@@ -3,6 +3,7 @@
 <?php
 
 require __DIR__ . '/vendor/autoload.php';
+require __DIR__ . '/lib/Global/Global.php';
 
 use Illuminate\Console\Application;
 use Illuminate\Container\Container;
@@ -10,16 +11,25 @@ use Illuminate\Events\Dispatcher;
 use Symfony\Component\Finder\Finder;
 
 $finder = new Finder();
-$finder->files()->in(__DIR__ . '/app/Console/Commands/')->name('*.php');
+$finder->files()->in([
+    sprintf('%s/Console/', lib_path()),
+    sprintf('%s/Console/Commands/', app_path()),
+])->name('*.php');
 
 $container = new Container();
 $dispatcher = new Dispatcher($container);
 $application = new Application($container, $dispatcher, '2.1.0');
 
+$namespaces = ['App\Console\Commands\\', 'Lib\Console\\'];
+
 foreach ($finder as $file) {
-    $commandClass = 'App\Console\Commands\\' . $file->getBasename('.php');
-    if (class_exists($commandClass)) {
-        $application->add(new $commandClass());
+    foreach ($namespaces as $namespace) {
+        $commandClass = $namespace . $file->getBasename('.php');
+
+        if (class_exists($commandClass)) {
+            $application->add(new $commandClass());
+            break;
+        }
     }
 }
