+++ lib/Console/CreateCommand.php
@@ -1,11 +1,18 @@
 <?php
 
-namespace App\Console\Commands;
+namespace Lib\Console;
 
-require_once __DIR__ . '/../../../lib/Global/Global.php';
+require_once __DIR__ . '/../Global/Global.php';
 
 use Illuminate\Console\Command;
 
+/**
+ * Class CreateCommand
+ * 
+ * This command will create a new command
+ * 
+ * @author Joshua McLean
+ */
 class CreateCommand extends Command
 {
     protected $signature = 'create:command {name}';
